Move(#42): AuthMemberService 패키지 변경
