+++ client/google-oauth-client/src/main/java/com/wypl/googleoauthclient/utils/AuthenticatedArgumentResolver.java
@@ -10,7 +10,7 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 import com.wypl.googleoauthclient.annotation.Authenticated;
-import com.wypl.googleoauthclient.data.AuthMemberService;
+import com.wypl.googleoauthclient.service.AuthMemberService;
 import com.wypl.googleoauthclient.domain.AuthMember;
 import com.wypl.googleoauthclient.exception.GoogleOAuthErrorCode;
 import com.wypl.googleoauthclient.exception.GoogleOAuthException;
